fixes for fastapi easy command
