+++ fast_app/services.py
@@ -1,10 +1,10 @@
-from services_modules.crawler import Crawler 
-from services_modules.scraper import Scraper
+from .services_modules.crawler import Crawler 
+from .services_modules.scraper import Scraper
 
 # The region parameter needs to be according to yahoo finances regions list
 region = 'Qatar'
 
-class ScrapreCrawlerService:
+class ScraperCrawlerService:
     def generate_stocks_data(self, region: str):
         crawler = Crawler()
         finance_table_html = crawler.crawl_finances(region)
