+++ fast_app/services_modules/scraper.py
