+++ requirements.txt
@@ -4,15 +4,18 @@ attrs==22.2.0
 beautifulsoup4==4.12.2
 certifi==2022.12.7
 charset-normalizer==3.1.0
+click==8.1.3
 exceptiongroup==1.1.1
 fastapi==0.95.0
 h11==0.14.0
+httptools==0.5.0
 idna==3.4
 outcome==1.2.0
 packaging==23.0
 pydantic==1.10.7
 PySocks==1.7.1
 python-dotenv==1.0.0
+PyYAML==6.0
 requests==2.28.2
 selenium==4.8.3
 sniffio==1.3.0
@@ -24,5 +27,9 @@ trio==0.22.0
 trio-websocket==0.10.2
 typing_extensions==4.5.0
 urllib3==1.26.15
+uvicorn==0.21.1
+uvloop==0.17.0
+watchfiles==0.19.0
 webdriver-manager==3.8.5
+websockets==11.0.1
 wsproto==1.2.0
