+++ run.sh
@@ -1,0 +1,1 @@
+uvicorn fast_app.main:app --reload --port 8000
